Track changes to task count & status in statusbar
